Backend of the time windows and items menu

+++ src/main/java/hu/gerviba/webschop/model/OrderEntity.java
@@ -49,7 +49,7 @@ public class OrderEntity implements Serializable {
     private String detailsJson;
 
     @Column
-    private int intervalId;
+    private long intervalId;
     
     @Column
     private String intervalMessage;
